requirements.txt
Appium-Python-Client
prettytable